requirements.txt
nltk==3.7
numpy==1.23.4
openai==0.25.0
torch==1.12.1
transformers==4.23.1
accelerate==0.13.2
mypy==0.982
lxml==4.9.1
tqdm-stubs==0.2.1
pylint==2.15.5
pylint-json2html==0.4.0
pylint-exit==1.2.0
pytest==7.2.0
pytest-html==3.2.0
coverage==6.5.0
black==22.10.0
ipython==8.5.0